+++ NEWS.md
@@ -1,3 +1,13 @@
+# pointblank 0.5.1
+
+* Improved compatibility with validations performed on SQL Server 2019.
+
+* Integrated the `label` argument into all validation functions; this label is available in the agent `x_list` and, more importantly, displayed in the agent report (in the `STEP` column).
+
+* Added the `"combined"` option in the `get_sundered_data()` function (for the `type` argument). This applies a categorical (pass/fail) label (settable in the new `pass_fail` argument of the same function) in a new `.pb_combined` flag column of the output table.
+
+* Made several visual improvements to the agent report.
+
 # pointblank 0.5.0
 
 ## New features
@@ -26,7 +36,7 @@
 
 * New **testthat** tests were added that test **pointblank** validations against mock PostgreSQL and MySQL database tables via the **dittodb** package. Thank you @pachamaltese for implementing these tests.
 
-# pointblank 0.4.0 (2020-06-22)
+# pointblank 0.4.0
 
 ## New R Markdown features
 
@@ -82,7 +92,7 @@
 
 * Added column validity checks inside of internal `interrogate_*()` functions
 
-# pointblank 0.3.1 (2020-04-02)
+# pointblank 0.3.1
 
 * Fixed implementation of the `col_vals_between()` and `col_vals_not_between()` step functions to work with `tbl_dbi` objects.
 
@@ -102,7 +112,7 @@
 
 * Numerous fixes to ensure compatibility with tibble 3.0.0
 
-# pointblank 0.3.0 (2020-01-10)
+# pointblank 0.3.0
 
 The pointblank package has been changed significantly from the previous version in favor of consistency and simplicity, better reporting, and increased power. The internals have been extensively refactored and the API has accordingly gone through revisions.
 
@@ -126,12 +136,12 @@ The pointblank package has been changed significantly from the previous version 
 
 * The `conjointly()` function is a new validation step function that allows for multiple rowwise validation steps to be performed for joint validity testing.
 
-# pointblank 0.2.1 (2019-09-12)
+# pointblank 0.2.1
 
 * Revisions on account of API changes in **tidyr** `1.0.0`.
 
 * Incorporates corrections related to API changes in **rlang** `0.2.0`.
 
-# pointblank 0.1 (2017-08-25)
+# pointblank 0.1
 
 * First release.